+++ Shared/Enums/Screens/CustomWatchListFilters.swift
@@ -1,0 +1,24 @@
+//
+//  CustomWatchListFilters.swift
+//  Story (iOS)
+//
+//  Created by Alexandre Madeira on 16/02/23.
+//
+
+import Foundation
+
+enum CustomWatchListFilters: String, Identifiable, CaseIterable {
+    var id: String { rawValue }
+    case all, movies, shows
+    
+    var localizableTitle: String {
+        switch self {
+        case .all:
+            return NSLocalizedString("allCustomWatchListFilters", comment: "")
+        case .movies:
+            return NSLocalizedString("moviesCustomWatchListFilters", comment: "")
+        case .shows:
+            return NSLocalizedString("showsCustomWatchListFilters", comment: "")
+        }
+    }
+}
